Check IMP modules after importing
A simple "import IMP.foo" isn't sufficient to check if the
IMP.foo module built and installed correctly, since Python 3
will let us import IMP.foo if a foo subdirectory exists (e.g.
left over from an older version of IMP). For a more rigorous
test, assert that the contents of the module are as expected.