Make sure that modXXX binary uses Apple's Python
The modXXX binary (e.g. mod9.21 for the current release)
is provided for compatibility with old Modeller tutorials
and to run TOP scripts. It is linked against Apple's Python
interpreter, so can fail if it tries to load in Homebrew's
newer Python libraries (which can happen on a 10.14 machine).
Set PYTHONHOME to force it to use Apple's Python libraries.
Closes #1.